+++ RELEASE-NOTES.md
@@ -37,6 +37,8 @@ to use Open Cascade's file I/O capabilities in support of Quest applications.
   of `loadExternalData` in `sidre::IOManager` and `sidre::Group`.
 
 ###  Changed
+- `primal::NumericArray` has been moved to `core`.  The header is
+  `core/utilities/NumericArray.hpp`.
 - `quest::Shaper` and `quest::IntersectionShaper` constructors require a runtime policy.
   Changing the policy after construction is no longer supported.
 - Importing Conduit array data into `sidre::View` now allocates destination